requirements.txt
customtkinter
opencv-python
tensorflow
pillow
numpy
scikit-learn